Add utils file / refactor vacation-form
